Renamed ApiPageHandler type to Handler
